Refactored IO into an interface
