+++ gbc/gbc.go
@@ -31,7 +31,7 @@ type GomeboyColor struct {
 	gpu          *gpu.GPU
 	cpu          *cpu.GbcCPU
 	mmu          *mmu.GbcMMU
-	io           *inputoutput.IO
+	io           inputoutput.IOHandler
 	apu          *apu.APU
 	timer        *timer.Timer
 	fpsCounter   *metric.FPSCounter
@@ -45,8 +45,8 @@ type GomeboyColor struct {
 	inBootMode   bool
 }
 
-func Init(cart *cartridge.Cartridge, saveStore saves.Store, conf *config.Config) (*GomeboyColor, error) {
-	var gbc *GomeboyColor = newGomeboyColor(cart, conf, saveStore)
+func Init(cart *cartridge.Cartridge, saveStore saves.Store, conf *config.Config, ioHandler inputoutput.IOHandler) (*GomeboyColor, error) {
+	var gbc *GomeboyColor = newGomeboyColor(cart, conf, saveStore, ioHandler)
 
 	b, er := gbc.mmu.LoadBIOS(BOOTROM)
 	if !b {
@@ -82,11 +82,11 @@ func Init(cart *cartridge.Cartridge, saveStore saves.Store, conf *config.Config)
 	}
 	defer r.Close()
 
-	gbc.gpu.LinkScreen(gbc.io.ScreenOutputChannel)
+	gbc.gpu.LinkScreen(gbc.io.GetScreenOutputChannel())
 
 	gbc.setupBoot()
 
-	err = gbc.io.Init(gbc.config.Title, gbc.config.ScreenSize, gbc.config.Headless, gbc.onClose)
+	err = gbc.io.Init(gbc.config.Title, gbc.config.ScreenSize, gbc.onClose)
 	if err != nil {
 		log.Fatalln("io init failure\n\t", err)
 	}
@@ -145,23 +145,22 @@ func (gbc *GomeboyColor) Reset() {
 	gbc.gpu.Reset()
 	gbc.mmu.Reset()
 	gbc.apu.Reset()
-	gbc.io.KeyHandler.Reset()
-	gbc.io.ScreenOutputChannel <- &(types.Screen{})
+	gbc.io.GetKeyHandler().Reset()
 	gbc.setupBoot()
 }
 
-func newGomeboyColor(cart *cartridge.Cartridge, conf *config.Config, saveStore saves.Store) *GomeboyColor {
+func newGomeboyColor(cart *cartridge.Cartridge, conf *config.Config, saveStore saves.Store, ioHandler inputoutput.IOHandler) *GomeboyColor {
 	gbc := new(GomeboyColor)
 
 	gbc.cart = cart
 	gbc.config = conf
 	gbc.saveStore = saveStore
+	gbc.io = ioHandler
 	gbc.debugOptions = new(DebugOptions)
 	gbc.fpsCounter = metric.NewFPSCounter()
 	gbc.mmu = mmu.NewGbcMMU()
 	gbc.cpu = cpu.NewCPU(gbc.mmu)
 
-	gbc.io = inputoutput.NewIO()
 	gbc.gpu = gpu.NewGPU()
 	gbc.apu = apu.NewAPU()
 	gbc.timer = timer.NewTimer()
@@ -169,14 +168,14 @@ func newGomeboyColor(cart *cartridge.Cartridge, conf *config.Config, saveStore s
 	//mmu will process interrupt requests from GPU (i.e. it will set appropriate flags)
 	gbc.gpu.LinkIRQHandler(gbc.mmu)
 	gbc.timer.LinkIRQHandler(gbc.mmu)
-	gbc.io.KeyHandler.LinkIRQHandler(gbc.mmu)
+	gbc.io.GetKeyHandler().LinkIRQHandler(gbc.mmu)
 
 	gbc.mmu.ConnectPeripheral(gbc.apu, 0xFF10, 0xFF3F)
 	gbc.mmu.ConnectPeripheral(gbc.gpu, 0x8000, 0x9FFF)
 	gbc.mmu.ConnectPeripheral(gbc.gpu, 0xFE00, 0xFE9F)
 	gbc.mmu.ConnectPeripheral(gbc.gpu, 0xFF57, 0xFF6F)
 	gbc.mmu.ConnectPeripheralOn(gbc.gpu, 0xFF40, 0xFF41, 0xFF42, 0xFF43, 0xFF44, 0xFF45, 0xFF47, 0xFF48, 0xFF49, 0xFF4A, 0xFF4B, 0xFF4F)
-	gbc.mmu.ConnectPeripheralOn(gbc.io.KeyHandler, 0xFF00)
+	gbc.mmu.ConnectPeripheralOn(gbc.io.GetKeyHandler(), 0xFF00)
 	gbc.mmu.ConnectPeripheralOn(gbc.timer, 0xFF04, 0xFF05, 0xFF06, 0xFF07)
 
 	return gbc
