+++ inputoutput/inputoutput.go
@@ -4,155 +4,34 @@ package inputoutput
 
 import (
 	"log"
-	"time"
 
-	"github.com/djhworld/gomeboycolor/components"
-	"github.com/djhworld/gomeboycolor/constants"
 	"github.com/djhworld/gomeboycolor/types"
 
 	"github.com/go-gl/gl/v2.1/gl"
 	"github.com/go-gl/glfw/v3.2/glfw"
 )
 
-const PREFIX string = "IO"
-const ROW_1 byte = 0x10
-const ROW_2 byte = 0x20
-const SCREEN_WIDTH int = 160
-const SCREEN_HEIGHT int = 144
-
 var DefaultControlScheme ControlScheme = ControlScheme{
-	glfw.KeyUp,
-	glfw.KeyDown,
-	glfw.KeyLeft,
-	glfw.KeyRight,
-	glfw.KeyZ,
-	glfw.KeyX,
-	glfw.KeyA,
-	glfw.KeyS,
+	int(glfw.KeyUp),
+	int(glfw.KeyDown),
+	int(glfw.KeyLeft),
+	int(glfw.KeyRight),
+	int(glfw.KeyZ),
+	int(glfw.KeyX),
+	int(glfw.KeyA),
+	int(glfw.KeyS),
 }
 
-type ControlScheme struct {
-	UP     glfw.Key
-	DOWN   glfw.Key
-	LEFT   glfw.Key
-	RIGHT  glfw.Key
-	A      glfw.Key
-	B      glfw.Key
-	START  glfw.Key
-	SELECT glfw.Key
+type GLFWBasedIO struct {
+	*CoreIO
 }
 
-type KeyHandler struct {
-	controlScheme ControlScheme
-	colSelect     byte
-	rows          [2]byte
-	irqHandler    components.IRQHandler
+func NewGLFWBasedIO(frameRateLock int64, headless bool) *GLFWBasedIO {
+	log.Println("Creating GLFW based IO Handler")
+	return &GLFWBasedIO{newCoreIO(frameRateLock, headless)}
 }
 
-func (k *KeyHandler) Init(cs ControlScheme) {
-	k.controlScheme = cs
-	k.Reset()
-}
-
-func (k *KeyHandler) Name() string {
-	return PREFIX + "-KEYB"
-}
-
-func (k *KeyHandler) Reset() {
-	log.Printf("%s: Resetting", k.Name())
-	k.rows[0], k.rows[1] = 0x0F, 0x0F
-	k.colSelect = 0x00
-}
-
-func (k *KeyHandler) LinkIRQHandler(m components.IRQHandler) {
-	k.irqHandler = m
-	log.Printf("%s: Linked IRQ Handler to Keyboard Handler", k.Name())
-}
-
-func (k *KeyHandler) Read(addr types.Word) byte {
-	var value byte
-
-	switch k.colSelect {
-	case ROW_1:
-		value = k.rows[1]
-	case ROW_2:
-		value = k.rows[0]
-	default:
-		value = 0x00
-	}
-
-	return value
-}
-
-func (k *KeyHandler) Write(addr types.Word, value byte) {
-	k.colSelect = value & 0x30
-}
-
-//released sets bit for key to 0
-func (k *KeyHandler) KeyDown(key glfw.Key) {
-	k.irqHandler.RequestInterrupt(constants.JOYP_HILO_IRQ)
-	switch key {
-	case k.controlScheme.UP:
-		k.rows[0] &= 0xB
-	case k.controlScheme.DOWN:
-		k.rows[0] &= 0x7
-	case k.controlScheme.LEFT:
-		k.rows[0] &= 0xD
-	case k.controlScheme.RIGHT:
-		k.rows[0] &= 0xE
-	case k.controlScheme.A:
-		k.rows[1] &= 0xE
-	case k.controlScheme.B:
-		k.rows[1] &= 0xD
-	case k.controlScheme.START:
-		k.rows[1] &= 0x7
-	case k.controlScheme.SELECT:
-		k.rows[1] &= 0xB
-	}
-}
-
-//released sets bit for key to 1
-func (k *KeyHandler) KeyUp(key glfw.Key) {
-	switch key {
-	case k.controlScheme.UP:
-		k.rows[0] |= 0x4
-	case k.controlScheme.DOWN:
-		k.rows[0] |= 0x8
-	case k.controlScheme.LEFT:
-		k.rows[0] |= 0x2
-	case k.controlScheme.RIGHT:
-		k.rows[0] |= 0x1
-	case k.controlScheme.A:
-		k.rows[1] |= 0x1
-	case k.controlScheme.B:
-		k.rows[1] |= 0x2
-	case k.controlScheme.START:
-		k.rows[1] |= 0x8
-	case k.controlScheme.SELECT:
-		k.rows[1] |= 0x4
-	}
-}
-
-type IO struct {
-	KeyHandler          *KeyHandler
-	Display             *Display
-	ScreenOutputChannel chan *types.Screen
-	AudioOutputChannel  chan int
-	headless            bool
-}
-
-func NewIO() *IO {
-	var i *IO = new(IO)
-	i.KeyHandler = new(KeyHandler)
-	i.Display = new(Display)
-	i.ScreenOutputChannel = make(chan *types.Screen)
-	i.AudioOutputChannel = make(chan int)
-	i.headless = false
-	return i
-}
-
-func (i *IO) Init(title string, screenSize int, headless bool, onCloseHandler func()) error {
-	i.headless = headless
+func (i *GLFWBasedIO) Init(title string, screenSize int, onCloseHandler func()) error {
 	var err error = nil
 
 	if !i.headless {
@@ -164,39 +43,19 @@ func (i *IO) Init(title string, screenSize int, headless bool, onCloseHandler fu
 		i.KeyHandler.Init(DefaultControlScheme) //TODO: allow user to define controlscheme
 		i.Display.window.SetKeyCallback(func(w *glfw.Window, key glfw.Key, scancode int, action glfw.Action, mods glfw.ModifierKey) {
 			if action == glfw.Repeat {
-				i.KeyHandler.KeyDown(key)
+				i.KeyHandler.KeyDown(int(key))
 				return
 			}
 
 			if action == glfw.Press {
-				i.KeyHandler.KeyDown(key)
+				i.KeyHandler.KeyDown(int(key))
 			} else {
-				i.KeyHandler.KeyUp(key)
+				i.KeyHandler.KeyUp(int(key))
 			}
 		})
 	}
 
 	return err
-}
-
-//This will wait for updates to the display or audio and dispatch them accordingly
-func (i *IO) Run() {
-	fpsLock := time.Second / 30.0
-	fpsThrottler := time.Tick(fpsLock)
-
-	for {
-		select {
-		case data := <-i.ScreenOutputChannel:
-			if !i.headless {
-				<-fpsThrottler
-				i.Display.drawFrame(data)
-			}
-		case data := <-i.AudioOutputChannel:
-			if !i.headless {
-				log.Println("Writing %d to audio!", data)
-			}
-		}
-	}
 }
 
 type Display struct {
